bitstream.c:	Table based CRC calculation code added (must be enabled to use) brhist.c:	Enlighted design of the bit rate history output main.c:		init of CRC Table in bitstream.c parse.c:	Some parameters can be name in Hz/bps and in kHz/kbps timestatus.c:	Enlighted design of the time counters lame.c:		Changed screen update frequency from 50/100 frames to 2 seconds

+++ timestatus.h
@@ -1,8 +1,10 @@
 #ifndef TIMESTATUS_H_INCLUDED
 #define TIMESTATUS_H_INCLUDED
 
-
-void timestatus(int samp_rate,long frameNum,long totalframes, int framesize);
+void timestatus ( unsigned long samp_rate, 
+                  unsigned long frameNum, 
+                  unsigned long totalframes, 
+                  int           framesize);
 void timestatus_finish(void);
 
 #if (defined LIBSNDFILE || defined LAMESNDFILE)
